requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
oauthlib==3.2.2
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
tweepy==4.14.0
urllib3==2.1.0
